ICP参数

+++ R360_StructureLight/inc/config.h
@@ -1,7 +1,7 @@
 
 #ifndef CONFIG_H
 #define CONFIG_H
-extern int config_num[50];
+extern float config_num[50];
 
 void Get_ConfigNum();
 void ArgvConfig();
